removed redundant classes and refactored ClockController into view package
