--- src/main/java/gui/view/DailyCalendarController.java
@@ -1,47 +1,0 @@
-package gui.view;
-
-import com.jfoenix.controls.JFXDrawer;
-import gui.utility.NavigationHelper;
-import gui.viewmodel.ViewModel;
-import javafx.event.ActionEvent;
-import javafx.fxml.FXML;
-import javafx.fxml.Initializable;
-import javafx.scene.control.ComboBox;
-import javafx.scene.layout.AnchorPane;
-
-import java.io.IOException;
-import java.net.URL;
-import java.util.ResourceBundle;
-
-public class DailyCalendarController implements Initializable, ViewModelBindingController {
-
-    @FXML
-    private JFXDrawer collapsedNavPanel;
-
-    @FXML
-    private JFXDrawer extendedNavPanel;
-
-    @FXML
-    private AnchorPane mainBackground;
-
-    @FXML
-    private ComboBox<String> calendarType;
-
-    @Override
-    public void initialize(URL location, ResourceBundle resources) {
-        NavigationHelper.initializeNavPanel(extendedNavPanel, collapsedNavPanel);
-
-        calendarType.getItems().addAll("Month", "Week", "Day");
-    }
-
-    @FXML
-    void calendarTypeSelected(ActionEvent event) throws IOException {
-        String selected = calendarType.getValue();
-        NavigationHelper.switchCalendarPageType(event, selected);
-    }
-
-    @Override
-    public void init(ViewModel viewModel) {
-
-    }
-}
