--- src/main/java/gui/viewmodel/NavigationCollapsedPanelViewModel.java
@@ -1,4 +1,0 @@
-package gui.viewmodel;
-
-public class NavigationCollapsedPanelViewModel {
-}
