--- src/main/java/gui/viewmodel/NavigationExtendedPanelViewModel.java
@@ -1,4 +1,0 @@
-package gui.viewmodel;
-
-public class NavigationExtendedPanelViewModel {
-}
